refactor: migrate models to entities
